move all mappings to semantic_mappings directory add README.md
