+++ chord_metadata_service/restapi/semantic_mappings/README.md
@@ -1,0 +1,6 @@
+# Semantic mappings
+
+This folder contains dictionaries for mapping phenopackets data to:
+- FHIR: based on [Phenopackets of FHIR Implementation guide](https://aehrc.github.io/fhir-phenopackets-ig/).
+Some of the mappings use [HL7 Genomics Reporting Implementation Guide](http://hl7.org/fhir/uv/genomics-reporting/STU1/index.html)
+- schema.org: schema.org context provided to create jsonld seriliazation of the dataset data.
